requirements.txt
pydub==0.25.1
openai==1.10.0
anthropic==0.16.0
python-dotenv==1.0.0
tqdm==4.66.1
PyMuPDF>=1.23.0
Pillow>=10.0.0 